requirements.txt
llama-cpp-python>=0.3.0
duckduckgo-search>=8.0.0
rich>=13.0.0
